Refactor random_infohash
