+++ src/tribler-core/tribler_core/components/gigachannel_manager/tests/test_gigachannel_manager.py
@@ -17,7 +17,7 @@ from tribler_core.components.gigachannel_manager.gigachannel_manager import Giga
 from tribler_core.components.metadata_store.db.orm_bindings.channel_node import NEW
 from tribler_core.tests.tools.base_test import MockObject
 from tribler_core.tests.tools.common import TORRENT_UBUNTU_FILE
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 
 update_metainfo = None
