+++ src/tribler-core/tribler_core/components/metadata_store/db/orm_bindings/channel_metadata.py
@@ -27,7 +27,7 @@ from tribler_core.components.metadata_store.db.serialization import (
     HealthItemsPayload,
 )
 from tribler_core.utilities.path_util import Path
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 from tribler_core.utilities.unicode import hexlify
 
 CHANNEL_DIR_NAME_PK_LENGTH = 32  # Its not 40 so it could be distinguished from infohash
