+++ src/tribler-core/tribler_core/components/metadata_store/db/tests/test_torrent_metadata.py
@@ -14,7 +14,7 @@ from tribler_core.components.metadata_store.db.orm_bindings.discrete_clock impor
 from tribler_core.components.metadata_store.db.orm_bindings.torrent_metadata import tdef_to_metadata_dict
 from tribler_core.components.metadata_store.db.serialization import CHANNEL_TORRENT, REGULAR_TORRENT
 from tribler_core.tests.tools.common import TORRENT_UBUNTU_FILE
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 
 EMPTY_BLOB = b""
