+++ src/tribler-core/tribler_core/components/metadata_store/remote_query_community/tests/test_remote_query_community.py
@@ -1,4 +1,5 @@
 import random
+import string
 from asyncio import sleep
 from binascii import unhexlify
 from datetime import datetime
@@ -7,8 +8,6 @@ from operator import attrgetter
 from os import urandom
 from time import time
 from unittest.mock import Mock, patch
-
-from faker import Faker
 
 from ipv8.keyvault.crypto import default_eccrypto
 from ipv8.test.base import TestBase
@@ -27,16 +26,14 @@ from tribler_core.components.metadata_store.remote_query_community.remote_query_
 )
 from tribler_core.components.metadata_store.remote_query_community.settings import RemoteQueryCommunitySettings
 from tribler_core.utilities.path_util import Path
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 from tribler_core.utilities.unicode import hexlify
 
 # pylint: disable=protected-access
 
-fake = Faker()
 
-
-def fake_sentence():
-    return fake.sentence(nb_words=4)
+def random_string():
+    return ''.join(random.choices(string.ascii_uppercase + string.digits, k=100))
 
 
 def add_random_torrent(metadata_cls, name="test", channel=None, seeders=None, leechers=None, last_check=None):
@@ -243,7 +240,7 @@ class TestRemoteQueryCommunity(TestBase):
         mds1 = self.nodes[1].overlay.mds
 
         with db_session:
-            chan = mds0.ChannelMetadata.create_channel(fake_sentence(), "")
+            chan = mds0.ChannelMetadata.create_channel(random_string(), "")
             torrent_infohash = random_infohash()
             torrent = mds0.TorrentMetadata(origin_id=chan.id_, infohash=torrent_infohash, title='title1')
             torrent.sign()
@@ -294,7 +291,7 @@ class TestRemoteQueryCommunity(TestBase):
 
         with db_session:
             for _ in range(0, 100):
-                mds0.ChannelMetadata.create_channel(fake_sentence(), "")
+                mds0.ChannelMetadata.create_channel(random_string(), "")
 
         peer = self.nodes[0].my_peer
         kwargs_dict = {"metadata_type": [CHANNEL_TORRENT]}
