+++ src/tribler-core/tribler_core/components/metadata_store/restapi/tests/test_channels_endpoint.py
@@ -24,7 +24,7 @@ from tribler_core.components.metadata_store.utils import RequestTimeoutException
 from tribler_core.components.restapi.rest.base_api_test import do_request
 from tribler_core.components.restapi.rest.rest_manager import error_middleware
 from tribler_core.tests.tools.common import TORRENT_UBUNTU_FILE
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 from tribler_core.utilities.unicode import hexlify
 
 PNG_DATA = unhexlify(
