+++ src/tribler-core/tribler_core/components/popularity/community/tests/test_popularity_community.py
@@ -17,7 +17,7 @@ from tribler_core.components.metadata_store.remote_query_community.settings impo
 from tribler_core.components.popularity.community.popularity_community import PopularityCommunity
 from tribler_core.tests.tools.base_test import MockObject
 from tribler_core.utilities.path_util import Path
-from tribler_core.utilities.random_utils import random_infohash
+from tribler_core.utilities.utilities import random_infohash
 
 
 class TestPopularityCommunity(TestBase):
