--- src/tribler-core/tribler_core/utilities/tests/test_random_utils.py
@@ -1,8 +1,0 @@
-from tribler_core.utilities.random_utils import random_infohash
-
-
-def test_random_infohash():
-    test_infohash = random_infohash()
-    assert test_infohash
-    assert isinstance(test_infohash, bytes)
-    assert len(test_infohash) == 20
